feat: Add updateOne method to UserController
Add the updateOne method to the UserController class in the users.js file. This method allows updating a user's last login date. It uses the userModel.update function to update the user's lastLogin field with the provided value. This change is necessary to track the last login time of users.